Rails authorization asset pipeline

+++ amazing_amazon/app/controllers/reviews_controller.rb
@@ -3,7 +3,7 @@ class ReviewsController < ApplicationController
 
     def create
         @product = Product.find params[:product_id]
-        @review = Review.new(params.require(:review).permit(:rating,:body))
+        @review = Review.new(params.require(:review).permit(:rating,:body,:is_hidden))
         @review.product = @product
         @review.user = @current_user
         
@@ -16,8 +16,16 @@ class ReviewsController < ApplicationController
 
     end
     def destroy
-        @product = Product.find params[:product_id]
-        @review = Review.find params[:id]
+      @product = Product.find params[:product_id]
+      @review = Review.find params[:id]
+      
+      # bellow if can? statment prevents a hacker from typing the destroy command in the url
+      # to see if passes.
+      if !(can? :delete, @review)
+        redirect_to @product, error: "not authorized"
+        # redirect_to product_path(@product), error: "not authorized" -- also works
+      end
+
         if @review.destroy
             flash[:success] = "Deleted"
             redirect_to product_path(@product)
@@ -26,4 +34,26 @@ class ReviewsController < ApplicationController
         end
 
     end
+
+    def unhide
+      @product = Product.find params[:product_id]
+      @review = Review.find params[:id]
+      
+      # bellow if can? statment prevents a hacker from typing the update command in the url
+      # to see if passes.
+      if !(can? :update, @review)
+        redirect_to @product, error: "not authorized"
+        # redirect_to product_path(@product), error: "not authorized" -- also works
+      end
+      
+      @review.is_hidden = false
+      
+      if @review.save
+          redirect_to product_path(@product.id)
+      else
+          redirect_to product_path(@product.id), error: 'review did not unhide!'
+      end
+
+  end
+
 end
